Issue #31 Добавлены команды mute и commandmute commandmute, кроме мута в чате, еще и запрещает выполнять команды

+++ src/main/java/org/ultramine/commands/CommandContext.java
@@ -423,7 +423,7 @@ public class CommandContext
 			return BasicTypeParser.parseItemStack(asString());
 		}
 		
-		public int asTimeMills()
+		public long asTimeMills()
 		{
 			return BasicTypeParser.parseTime(asString());
 		}
